+++ .storybook/components/LocationSearch/LocationSearch.stories.tsx
@@ -1,14 +1,14 @@
 import { BASE_HEADER_SCREEN_OPTIONS } from "@components/Navigation"
 import { TiFQueryClientProvider } from "@components/TiFQueryClientProvider"
-import { UserLocationFunctionsProvider } from "@hooks/UserLocation"
+import { UserLocationFunctionsProvider } from "@location"
 import { ArrayUtils } from "@lib/Array"
 import { delayData } from "@lib/DelayData"
 import {
   LocationSearchPicker,
-  useLocationSearchPicker
-} from "@location-search/Picker"
-import { LocationSearchBar } from "@location-search/SearchBar"
-import { mockLocationSearchResult } from "@location-search/mocks"
+  useLocationSearchPicker,
+  LocationSearchBar
+} from "@location/search"
+import { mockLocationSearchResult } from "@location/search/MockData"
 import {
   NavigationContainer,
   NavigationProp,
