+++ components/eventForm/EventFormValues.ts
@@ -1,6 +1,6 @@
 import { FixedDateRange } from "@date-time"
 import { EventColors, SaveEventRequest } from "@lib/events"
-import { LocationCoordinate2D } from "@lib/location"
+import { LocationCoordinate2D } from "@location/index"
 
 export type EventFormPlacemarkInfo = {
   readonly name?: string
