+++ components/eventForm/LocationBanner.tsx
@@ -1,5 +1,5 @@
-import { useReverseGeocodeQuery } from "../../hooks/Geocoding"
-import { placemarkToFormattedAddress } from "../../lib/location"
+import { useReverseGeocodeQuery } from "../../location/Geocoding"
+import { placemarkToFormattedAddress } from "../../location"
 import React from "react"
 import { StyleSheet, View } from "react-native"
 import { useEventFormContext } from "./EventForm"
