+++ event-details/arrival-tracking/MockData.ts
@@ -1,0 +1,8 @@
+import { faker } from "@faker-js/faker"
+import { mockLocationCoordinate2D } from "@location/MockData"
+import { EventArrival } from "./Models"
+
+export const mockEventArrival = (): EventArrival => ({
+  eventId: parseInt(faker.random.numeric(3)),
+  coordinate: mockLocationCoordinate2D()
+})
