+++ event-details/arrival-tracking/Refresh.ts
@@ -2,7 +2,7 @@ import { AsyncStorageUtils } from "@lib/AsyncStorage"
 import { StringDateSchema, addSecondsToDate, diffDates, now } from "@date-time"
 import AsyncStorage from "@react-native-async-storage/async-storage"
 import { AppState } from "react-native"
-import { EventArrival } from "./models"
+import { EventArrival } from "./Models"
 import { EventArrivalsTracker } from "./Tracker"
 
 /**
