+++ event-details/arrival-tracking/index.ts
@@ -1,4 +1,5 @@
 export * from "./Tracker"
-export * from "./models"
+export * from "./Models"
 export * from "./Geofencing"
 export * from "./UpcomingArrivals"
+export * from "./Refresh"
