+++ location/index.ts
@@ -1,4 +1,5 @@
-export * from "./ExpoHelpers"
 export * from "./Location"
 export * from "./Placemark"
 export * from "./Region"
+export * from "./Geocoding"
+export * from "./UserLocation"
