+++ location/search/Picker.tsx
@@ -1,8 +1,8 @@
 import { Headline } from "@components/Text"
 import { CircularIonicon } from "@components/common/Icons"
-import { useUserCoordinatesQuery } from "@hooks/UserLocation"
+import { useUserCoordinatesQuery } from "@location/UserLocation"
 import { AppStyles } from "@lib/AppColorStyle"
-import { LocationCoordinate2D, TiFLocation } from "@lib/location"
+import { LocationCoordinate2D, TiFLocation } from "@location"
 import { UseQueryResult, useQuery } from "@tanstack/react-query"
 import { LocationAccuracy, LocationObject } from "expo-location"
 import React from "react"
@@ -19,7 +19,7 @@ import {
   LocationSearchResultView
 } from "./SearchResultView"
 import { LocationSearchResultsListView } from "./SearchResultsList"
-import { LocationSearchResultsData } from "./models"
+import { LocationSearchResultsData } from "./Models"
 import { useLocationsSearchQueryObject } from "./state"
 
 export type UseLocationSearchPickerEnvironment = {
