+++ location/search/index.ts
@@ -1,7 +1,6 @@
 export * from "./Picker"
 export * from "./SearchResultsList"
 export * from "./SearchResultView"
-export * from "./models"
+export * from "./Models"
 export * from "./SearchBar"
 export * from "./RecentsStorage"
-
